backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
webauthn==2.2.0
python-jose[cryptography]==3.3.0
pydantic[email]==2.9.0
